Restructure ServerTime for better encapsulation
