+++ app/src/main/java/io/islnd/android/islnd/app/adapters/PostAdapter.java
@@ -27,7 +27,6 @@ import io.islnd.android.islnd.app.models.Profile;
 import io.islnd.android.islnd.app.util.ImageUtil;
 import io.islnd.android.islnd.app.util.Util;
 import io.islnd.android.islnd.app.viewholders.GlancePostViewHolder;
-import io.islnd.android.islnd.messaging.ServerTime;
 
 import java.util.ArrayList;
 
@@ -59,7 +58,7 @@ public class PostAdapter extends CursorRecyclerViewAdapter<GlancePostViewHolder>
         );
 
         holder.postUserName.setText(post.getUserName());
-        holder.postTimestamp.setText(Util.smartTimestampFromUnixTime(ServerTime.toLocalTimeMillis(post.getTimestamp())));
+        holder.postTimestamp.setText(Util.smartTimestampFromUnixTime(mContext, post.getTimestamp()));
         holder.postContent.setText(post.getContent());
         //TODO: Get actual comment count
         holder.postCommentCount.setText(Util.numberOfCommentsString(0));
