+++ app/src/main/java/io/islnd/android/islnd/messaging/event/EventListBuilder.java
@@ -14,6 +14,7 @@ import io.islnd.android.islnd.app.database.DataUtils;
 import io.islnd.android.islnd.app.util.ImageUtil;
 import io.islnd.android.islnd.app.util.Util;
 import io.islnd.android.islnd.messaging.ContentUtil;
+import io.islnd.android.islnd.messaging.ServerTime;
 
 public class EventListBuilder {
 
@@ -47,7 +48,7 @@ public class EventListBuilder {
                         getNewEventId(),
                         postId,
                         postText,
-                        ContentUtil.getContentTimestamp()));
+                        ServerTime.getCurrentTimeMillis(mContext)));
 
         return this;
     }
